courses and users routes
